codebase/requirements.txt
# Core
pydantic>=2.0

# API clients (install the ones you need)
openai>=1.0
anthropic>=0.30
google-generativeai>=0.5
huggingface_hub>=0.28
datasets>=3.0

# Inspect AI evaluation framework
inspect-ai>=0.3

# Analysis & visualisation
matplotlib>=3.7
numpy>=1.24
pandas>=2.0
seaborn>=0.13
scipy>=1.11

# Utilities
python-dotenv>=1.0
